+++ src/jdk.jfr/share/classes/jdk/jfr/Recording.java
@@ -38,7 +38,7 @@ import java.util.Objects;
 import jdk.jfr.internal.PlatformRecorder;
 import jdk.jfr.internal.PlatformRecording;
 import jdk.jfr.internal.Type;
-import jdk.jfr.internal.Utils;
+import jdk.jfr.internal.util.Utils;
 import jdk.jfr.internal.WriteableUserPath;
 
 /**
