+++ src/jdk.jfr/share/classes/jdk/jfr/consumer/EventStream.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -36,7 +36,6 @@ import java.util.Objects;
 import java.util.function.Consumer;
 
 import jdk.jfr.internal.SecuritySupport;
-import jdk.jfr.internal.Utils;
 import jdk.jfr.internal.consumer.EventDirectoryStream;
 import jdk.jfr.internal.consumer.EventFileStream;
 import jdk.jfr.internal.consumer.FileAccess;
@@ -120,7 +119,7 @@ public interface EventStream extends AutoCloseable {
      */
     @SuppressWarnings("removal")
     public static EventStream openRepository() throws IOException {
-        Utils.checkAccessFlightRecorder();
+        SecuritySupport.checkAccessFlightRecorder();
         return new EventDirectoryStream(
             AccessController.getContext(),
             null,
