+++ src/jdk.jfr/share/classes/jdk/jfr/internal/Control.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -35,6 +35,9 @@ import java.util.Set;
 
 import jdk.jfr.SettingControl;
 import jdk.jfr.internal.settings.JDKSettingControl;
+import jdk.jfr.internal.settings.PeriodSetting;
+import jdk.jfr.internal.settings.StackTraceSetting;
+import jdk.jfr.internal.settings.ThresholdSetting;
 
 final class Control {
     @SuppressWarnings("removal")
@@ -175,5 +178,18 @@ final class Control {
 
     final SettingControl getSettingControl() {
         return delegate;
+    }
+
+    boolean isVisible(boolean hasEventHook) {
+        if (isType(ThresholdSetting.class)) {
+            return !hasEventHook;
+        }
+        if (isType(PeriodSetting.class)) {
+            return hasEventHook;
+        }
+        if (isType(StackTraceSetting.class)) {
+            return !hasEventHook;
+        }
+        return true;
     }
 }
