+++ src/jdk.jfr/share/classes/jdk/jfr/internal/JVMSupport.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -25,6 +25,13 @@
 package jdk.jfr.internal;
 
 import java.io.IOException;
+import java.time.Instant;
+import java.time.LocalDateTime;
+
+import jdk.jfr.Recording;
+import jdk.jfr.internal.event.EventConfiguration;
+import jdk.jfr.internal.util.Utils;
+import jdk.jfr.internal.util.ValueFormatter;
 
 /**
  * Checks if the running VM supports Flight Recorder.
@@ -36,9 +43,13 @@ import java.io.IOException;
  * This is needed so a disabled-jfr.jar can be built for non Oracle JDKs.
  */
 public final class JVMSupport {
-
     private static final String UNSUPPORTED_VM_MESSAGE = "Flight Recorder is not supported on this VM";
     private static final boolean notAvailable = !checkAvailability();
+    /*
+     * This field will be lazily initialized and the access is not synchronized.
+     * The possible data race is benign and is worth of not introducing any contention here.
+     */
+    private static Instant lastTimestamp;
 
     private static boolean checkAvailability() {
         // set jfr.unsupported.vm to true to test API on an unsupported VM
@@ -83,5 +94,69 @@ public final class JVMSupport {
     }
 
     public static void tryToInitializeJVM() {
+    }
+
+    static long nanosToTicks(long nanos) {
+        return (long) (nanos * JVM.getJVM().getTimeConversionFactor());
+    }
+
+    static long getChunkStartNanos() {
+        long nanos = JVM.getJVM().getChunkStartNanos();
+        // JVM::getChunkStartNanos() may return a bumped timestamp, +1 ns or +2 ns.
+        // Spin here to give Instant.now() a chance to catch up.
+        awaitUniqueTimestamp();
+        return nanos;
+    }
+
+    private static void awaitUniqueTimestamp() {
+        if (lastTimestamp == null) {
+            lastTimestamp = Instant.now(); // lazy initialization
+        }
+        while (true) {
+            Instant time = Instant.now();
+            if (!time.equals(lastTimestamp)) {
+                lastTimestamp = time;
+                return;
+            }
+            try {
+                Thread.sleep(0, 100);
+            } catch (InterruptedException iex) {
+                // ignore
+            }
+        }
+    }
+
+    public static synchronized EventConfiguration getConfiguration(Class<? extends jdk.internal.event.Event> eventClass) {
+        Utils.ensureValidEventSubclass(eventClass);
+        Object configuration = JVM.getJVM().getConfiguration(eventClass);
+        if (configuration == null || configuration instanceof EventConfiguration) {
+            return (EventConfiguration) configuration;
+        }
+        throw new InternalError("Could not get configuration object on event class " + eventClass.getName());
+    }
+
+    public static synchronized void setConfiguration(Class<? extends jdk.internal.event.Event> eventClass, EventConfiguration configuration) {
+        Utils.ensureValidEventSubclass(eventClass);
+        if (!JVM.getJVM().setConfiguration(eventClass, configuration)) {
+            throw new InternalError("Could not set configuration object on event class " + eventClass.getName());
+        }
+    }
+
+    static boolean shouldInstrument(boolean isJDK, String name) {
+        if (!isJDK) {
+            return true;
+        }
+        if (!name.contains(".Container")) {
+            // Didn't match @Name("jdk.jfr.Container*") or class name "jdk.jfr.events.Container*"
+            return true;
+        }
+        return JVM.getJVM().isContainerized();
+    }
+
+    public static String makeFilename(Recording recording) {
+        String pid = JVM.getJVM().getPid();
+        String date = ValueFormatter.formatDateTime(LocalDateTime.now());
+        String idText = recording == null ? "" :  "-id-" + Long.toString(recording.getId());
+        return "hotspot-" + "pid-" + pid + idText + "-" + date + ".jfr";
     }
 }
