+++ src/jdk.jfr/share/classes/jdk/jfr/internal/MetadataReader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -51,6 +51,7 @@ import jdk.jfr.ValueDescriptor;
 import jdk.jfr.internal.MetadataDescriptor.Element;
 import jdk.jfr.internal.consumer.RecordingInput;
 import jdk.jfr.internal.consumer.StringParser;
+import jdk.jfr.internal.util.Utils;
 
 /**
  * Parses metadata.
