+++ src/jdk.jfr/share/classes/jdk/jfr/internal/MetadataRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -49,6 +49,7 @@ import jdk.jfr.ValueDescriptor;
 import jdk.jfr.internal.consumer.RepositoryFiles;
 import jdk.jfr.internal.event.EventConfiguration;
 import jdk.jfr.internal.periodic.PeriodicEvents;
+import jdk.jfr.internal.util.Utils;
 
 public final class MetadataRepository {
 
@@ -121,7 +122,7 @@ public final class MetadataRepository {
     }
 
     public synchronized void unregister(Class<? extends Event> eventClass) {
-        Utils.checkRegisterPermission();
+        SecuritySupport.checkRegisterPermission();
         EventConfiguration configuration = getConfiguration(eventClass, false);
         if (configuration != null) {
             configuration.getPlatformEventType().setRegistered(false);
@@ -133,7 +134,7 @@ public final class MetadataRepository {
     }
 
     public synchronized EventType register(Class<? extends jdk.internal.event.Event> eventClass, List<AnnotationElement> dynamicAnnotations, List<ValueDescriptor> dynamicFields) {
-        Utils.checkRegisterPermission();
+        SecuritySupport.checkRegisterPermission();
         if (jvm.isExcluded(eventClass)) {
             // Event classes are marked as excluded during class load
             // if they override methods in the jdk.jfr.Event class, i.e. commit().
@@ -181,7 +182,7 @@ public final class MetadataRepository {
         if (ensureInitialized) {
             Utils.ensureInitialized(eventClass);
         }
-        return Utils.getConfiguration(eventClass);
+        return JVMSupport.getConfiguration(eventClass);
     }
 
     private EventConfiguration newEventConfiguration(EventType eventType, EventControl ec) {
@@ -209,10 +210,10 @@ public final class MetadataRepository {
         PlatformEventType pe = configuration.getPlatformEventType();
         pe.setRegistered(true);
         // If class is instrumented or should not be instrumented, mark as instrumented.
-        if (jvm.isInstrumented(eventClass) || !Utils.shouldInstrument(pe.isJDK(), pe.getName())) {
+        if (jvm.isInstrumented(eventClass) || !JVMSupport.shouldInstrument(pe.isJDK(), pe.getName())) {
             pe.setInstrumented();
         }
-        Utils.setConfiguration(eventClass, configuration);
+        JVMSupport.setConfiguration(eventClass, configuration);
         return configuration;
     }
 
@@ -231,7 +232,7 @@ public final class MetadataRepository {
         ArrayList<EventControl> controls = new ArrayList<>(eventClasses.size() + nativeControls.size());
         controls.addAll(nativeControls);
         for (Class<? extends jdk.internal.event.Event> clazz : eventClasses) {
-            EventConfiguration eh = Utils.getConfiguration(clazz);
+            EventConfiguration eh = JVMSupport.getConfiguration(clazz);
             if (eh != null) {
                 controls.add(eh.getEventControl());
             }
@@ -248,7 +249,7 @@ public final class MetadataRepository {
         List<Class<? extends jdk.internal.event.Event>> allEventClasses = jvm.getAllEventClasses();
         List<EventConfiguration> eventConfigurations = new ArrayList<>(allEventClasses.size());
         for (Class<? extends jdk.internal.event.Event> clazz : allEventClasses) {
-            EventConfiguration ec = Utils.getConfiguration(clazz);
+            EventConfiguration ec = JVMSupport.getConfiguration(clazz);
             if (ec != null) {
                 eventConfigurations.add(ec);
             }
@@ -297,7 +298,7 @@ public final class MetadataRepository {
         // if the clock resolution is low, two chunks may
         // get the same timestamp. Utils.getChunkStartNanos()
         // ensures the timestamp is unique for the next chunk
-        long chunkStart = Utils.getChunkStartNanos();
+        long chunkStart = JVMSupport.getChunkStartNanos();
         if (filename != null) {
             RepositoryFiles.notifyNewFile();
         }
