+++ src/jdk.jfr/share/classes/jdk/jfr/internal/PrivateAccess.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -60,7 +60,7 @@ public abstract class PrivateAccess {
             // Will trigger
             // FlightRecorderPermission.<clinit>
             // which will call PrivateAccess.setPrivateAccess
-            new FlightRecorderPermission(Utils.REGISTER_EVENT);
+            new FlightRecorderPermission(SecuritySupport.REGISTER_EVENT);
         }
         return instance;
     }
