+++ src/jdk.jfr/share/classes/jdk/jfr/internal/SettingsManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -38,6 +38,7 @@ import java.util.StringJoiner;
 
 import jdk.jfr.internal.EventControl.NamedControl;
 import jdk.jfr.internal.event.EventConfiguration;
+import jdk.jfr.internal.util.Utils;
 
 final class SettingsManager {
 
@@ -155,7 +156,7 @@ final class SettingsManager {
     public void updateRetransform(List<Class<? extends jdk.internal.event.Event>> eventClasses) {
         List<Class<?>> classes = new ArrayList<>();
         for(Class<? extends jdk.internal.event.Event> eventClass: eventClasses) {
-            EventConfiguration ec = Utils.getConfiguration(eventClass);
+            EventConfiguration ec = JVMSupport.getConfiguration(eventClass);
             if (ec != null ) {
                 PlatformEventType eventType = ec.getPlatformEventType();
                 if (eventType.isMarkedForInstrumentation()) {
@@ -229,7 +230,7 @@ final class SettingsManager {
                 control.apply(values);
                 String after = control.getLastValue();
                 if (shouldLog) {
-                    if (Utils.isSettingVisible(control, ec.getEventType().hasEventHook())) {
+                    if (control.isVisible(ec.getEventType().hasEventHook())) {
                         if (values.size() > 1) {
                             StringJoiner sj = new StringJoiner(", ", "{", "}");
                             for (String s : values) {
