+++ src/jdk.jfr/share/classes/jdk/jfr/internal/ShutdownHook.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -32,6 +32,7 @@ import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
 
 import jdk.jfr.RecordingState;
+import jdk.jfr.internal.util.Utils;
 
 /**
  * Class responsible for dumping recordings on exit
@@ -80,7 +81,7 @@ final class ShutdownHook implements Runnable {
     @SuppressWarnings("removal")
     private WriteableUserPath makeDumpOnExitPath(PlatformRecording recording) {
         try {
-            String name = Utils.makeFilename(recording.getRecording());
+            String name = JVMSupport.makeFilename(recording.getRecording());
             AccessControlContext acc = recording.getNoDestinationDumpOnExitAccessControlContext();
             return AccessController.doPrivileged(new PrivilegedExceptionAction<WriteableUserPath>() {
                 @Override
