+++ src/jdk.jfr/share/classes/jdk/jfr/internal/Type.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -69,7 +69,7 @@ public class Type implements Comparable<Type> {
     static final Type CLASS = createKnownType(Class.class);
     static final Type STRING = createKnownType(String.class);
     static final Type THREAD = createKnownType(Thread.class);
-    static final Type STACK_TRACE = createKnownType(TYPES_PREFIX + "StackTrace", null);
+    public static final Type STACK_TRACE = createKnownType(TYPES_PREFIX + "StackTrace", null);
 
     private static Type createKnownType(Class<?> clazz) {
         return createKnownType(clazz.getName(), clazz);
