+++ src/jdk.jfr/share/classes/jdk/jfr/internal/dcmd/ArgumentParser.java
@@ -31,6 +31,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.StringJoiner;
+import jdk.jfr.internal.util.SpellChecker;
 
 final class ArgumentParser {
     private final Map<String, Object> options = new HashMap<>();
@@ -351,61 +352,17 @@ final class ArgumentParser {
     }
 
     void checkSpelling(Set<String> excludeSet) {
+        List<String> alternatives = new ArrayList<>();
+        for (Argument a : arguments) {
+            alternatives.add(a.name());
+        }
         for (String name : extendedOptions.keySet()) {
             if (!excludeSet.contains(name)) { // ignore names specified in .jfc
-                checkSpellingError(name);
-            }
-        }
-    }
-
-    private void checkSpellingError(String name) {
-        for (Argument a : arguments) {
-            String expected = a.name();
-            String s = name.toLowerCase();
-            int lengthDifference = expected.length() - s.length();
-            boolean spellingError = false;
-            if (lengthDifference == 0) {
-                if (expected.equals(s)) {
-                    spellingError = true; // incorrect case, or we wouldn't be here
-                } else {
-                    if (s.length() < 6) {
-                        spellingError = diff(expected, s) < 2; // one incorrect letter
-                    } else {
-                        spellingError = diff(expected, s) < 3; // two incorrect letter
-                    }
+                String suggestion = SpellChecker.check(name, alternatives);
+                if (suggestion != null) {
+                    throw new IllegalArgumentException("Error! Did you mean '" + suggestion + "' instead of '" + name + "'?");
                 }
             }
-            if (lengthDifference == 1) {
-                spellingError = inSequence(expected, s); // missing letter
-            }
-            if (lengthDifference == -1) {
-                spellingError = inSequence(s, expected); // additional letter
-            }
-            if (spellingError) {
-                throw new IllegalArgumentException("Error! Did you mean '" + expected + "' instead of '" + name + "'?");
-            }
         }
-    }
-
-    private int diff(String a, String b) {
-        int count = a.length();
-        for (int i = 0; i < a.length(); i++) {
-            if (a.charAt(i) == b.charAt(i)) {
-                count--;
-            }
-        }
-        return count;
-    }
-
-    private boolean inSequence(String longer, String shorter) {
-        int l = 0;
-        int s = 0;
-        while (l < longer.length() && s < shorter.length()) {
-            if (longer.charAt(l) == shorter.charAt(s)) {
-                s++;
-            }
-            l++;
-        }
-        return shorter.length() == s; // if 0, all letters in longer found in shorter
     }
 }
