+++ src/jdk.jfr/share/classes/jdk/jfr/internal/settings/ThresholdSetting.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -35,7 +35,7 @@ import jdk.jfr.Name;
 import jdk.jfr.Timespan;
 import jdk.jfr.internal.PlatformEventType;
 import jdk.jfr.internal.Type;
-import jdk.jfr.internal.Utils;
+import jdk.jfr.internal.util.ValueParser;
 @MetadataDefinition
 @Label("Threshold")
 @Name(Type.SETTINGS_PREFIX + "Threshold")
@@ -55,7 +55,7 @@ public final class ThresholdSetting extends JDKSettingControl {
         Long min = null;
         String text = null;
         for (String value : values) {
-            long l = Utils.parseTimespanWithInfinity(value);
+            long l = ValueParser.parseTimespanWithInfinity(value);
             // always accept first value
             if (min == null) {
                 min = l;
@@ -72,7 +72,7 @@ public final class ThresholdSetting extends JDKSettingControl {
 
     @Override
     public void setValue(String value) {
-        long l = Utils.parseTimespanWithInfinity(value);
+        long l = ValueParser.parseTimespanWithInfinity(value);
         this.value = value;
         eventType.setThreshold(l);
     }
