+++ test/jdk/jdk/jfr/jvm/TestFormatDuration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,7 +22,7 @@
  */
 package jdk.jfr.jvm;
 
-import jdk.jfr.internal.Utils;
+import jdk.jfr.internal.util.ValueFormatter;
 import jdk.test.lib.Asserts;
 
 import java.time.Duration;
@@ -33,7 +33,7 @@ import java.util.Locale;
  * @key jfr
  * @requires vm.hasJFR
  * @library /test/lib
- * @modules jdk.jfr/jdk.jfr.internal
+ * @modules jdk.jfr/jdk.jfr.internal.util
  * @run main/othervm jdk.jfr.jvm.TestFormatDuration
  */
 public class TestFormatDuration {
@@ -133,9 +133,9 @@ public class TestFormatDuration {
     private static void assertDuration(String value, String expected) throws Exception {
         long nanos = parse(value);
         System.out.println(value + " == " + expected + " ? (" + nanos + " ns) ");
-        Asserts.assertEquals(Utils.formatDuration(Duration.ofNanos(nanos)), expected);
+        Asserts.assertEquals(ValueFormatter.formatDuration(Duration.ofNanos(nanos)), expected);
         if (nanos != 0) {
-            Asserts.assertEquals(Utils.formatDuration(Duration.ofNanos(-nanos)), "-" + expected);
+            Asserts.assertEquals(ValueFormatter.formatDuration(Duration.ofNanos(-nanos)), "-" + expected);
         }
     }
